+++ Plugins/SPARK_PL/src/RenderingAPIs/PixelLight/SPK_PLQuadRendererFixedFunctions.cpp
@@ -143,7 +143,7 @@ void SPK_PLQuadRendererFixedFunctions::render(const SPK::Group &group)
 			pFixedFunctions->SetVertexBuffer(m_pSPK_PLBuffer->GetVertexBuffer());
 
 			// Draw
-			GetPLRenderer().DrawIndexedPrimitives(Primitive::TriangleList, 0, group.getNbParticles()*NumOfVerticesPerParticle-1, 0, group.getNbParticles()*NumOfIndicesPerParticle);
+			GetPLRenderer().DrawIndexedPrimitives(Primitive::TriangleList, 0, static_cast<uint32>(group.getNbParticles()*NumOfVerticesPerParticle-1), 0, static_cast<uint32>(group.getNbParticles()*NumOfIndicesPerParticle));
 		}
 	}
 }
